+++ crates/sequencer/src/serde.rs
@@ -1,4 +1,4 @@
-use std::{io, path::Path};
+use std::{fs, io, path::Path};
 
 use crate::state::State;
 use blockifier::{
@@ -19,6 +19,29 @@ pub trait DumpLoad {
     fn load_state_from_file(file_path: &Path) -> Result<Self, SerializationError>
     where
         Self: Sized;
+}
+
+impl DumpLoad for State {
+    /// This will serialize the current state, and will save it to a path
+    fn dump_state_to_file(self, path: &Path) -> Result<(), SerializationError> {
+        let serializable_state: SerializableState = self.into();
+
+        let dump = serde_json::to_string(&serializable_state)
+            .map_err(SerializationError::SerdeJsonError)?;
+
+        fs::write(path, dump).map_err(SerializationError::IoError)?;
+
+        Ok(())
+    }
+
+    /// This will read a dump from a file and initialize the state from it
+    fn load_state_from_file(path: &Path) -> Result<Self, SerializationError> {
+        let dump = fs::read(path).unwrap();
+        let serializable_state: SerializableState =
+            serde_json::from_slice(&dump).map_err(SerializationError::SerdeJsonError)?;
+
+        Ok(serializable_state.into())
+    }
 }
 
 #[derive(Error, Debug)]
@@ -47,6 +70,18 @@ impl From<State> for SerializableState {
             contracts: state.contracts,
             storage: state.storage,
             nonces: state.nonces,
+        }
+    }
+}
+
+impl From<SerializableState> for State {
+    fn from(serializable_state: SerializableState) -> Self {
+        Self {
+            classes: serializable_state.classes.clone(),
+            compiled_class_hashes: serializable_state.compiled_classes_hash.clone(),
+            contracts: serializable_state.contracts.clone(),
+            storage: serializable_state.storage.clone(),
+            nonces: serializable_state.nonces,
         }
     }
 }
@@ -96,9 +131,58 @@ mod serialize_contract_storage {
                         key_str, error
                     ))
                 })?;
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use blockifier::{
+        execution::contract_class::{ContractClass, ContractClassV0},
+        state::cached_state::ContractStorageKey,
+    };
 
-            map.insert(contract_storage_key, value_str.to_owned());
-        }
-        Ok(map)
+    use crate::{
+        constants::test_constants::{
+            ONE_CLASS_HASH, ONE_COMPILED_CLASS_HASH, ONE_FELT, TEST_CONTRACT, TEST_NONCE,
+            TEST_STORAGE_KEY,
+        },
+        state::State,
+    };
+
+    #[test]
+    pub fn dump_and_load_state() {
+        let mut state = State::default();
+
+        // setting up entry for state.classes
+        let class_hash = *ONE_CLASS_HASH;
+        let contract_class = include_str!("./test_data/cairo_0/compiled_classes/counter.json");
+        let contract_class: ContractClassV0 =  serde_json::from_str(contract_class).expect("failed to deserialize ContractClass from ./crates/sequencer/test_data/cairo_1/compiled_classes/account.json");
+        let contract_class = ContractClass::V0(contract_class);
+
+        let compiled_class_hash = *ONE_COMPILED_CLASS_HASH;
+        let contract_address = *TEST_CONTRACT;
+        let contract_storage_key: ContractStorageKey = (contract_address, *TEST_STORAGE_KEY);
+        let storage_value = *ONE_FELT;
+        let nonce = *TEST_NONCE;
+
+        state.classes.insert(class_hash, contract_class);
+        state
+            .compiled_class_hashes
+            .insert(class_hash, compiled_class_hash);
+        state.contracts.insert(contract_address, class_hash);
+        state.storage.insert(contract_storage_key, storage_value);
+        state.nonces.insert(contract_address, nonce);
+
+        let temp_file = tempfile::NamedTempFile::new().expect("failed open named temp file");
+        let dump_file_path = temp_file.into_temp_path();
+
+        state
+            .clone()
+            .dump_state_to_file(&dump_file_path)
+            .expect("failed to save dump to file");
+
+        let loaded_state =
+            State::load_state_from_file(&dump_file_path).expect("failed to load state from file");
+        assert_eq!(state, loaded_state);
+
+        dump_file_path.close().expect("failed to close temp file");
     }
 }
