+++ crates/sequencer/src/state.rs
@@ -1,6 +1,3 @@
-use std::fs;
-use std::path::Path;
-
 use blockifier::state::cached_state::CommitmentStateDiff;
 use blockifier::state::errors::StateError;
 use blockifier::state::state_api::{
@@ -20,7 +17,6 @@ use starknet_api::{
 use serde::{Deserialize, Serialize};
 
 use crate::commit::Committer;
-use crate::serde::{DumpLoad, SerializableState, SerializationError};
 
 /// Generic state structure for the sequencer.
 /// The use of `FxHashMap` allows for a better performance.
@@ -28,7 +24,7 @@ use crate::serde::{DumpLoad, SerializableState, SerializationError};
 /// which is faster than the default hash function. Think about changing
 /// if the test sequencer is used for tests outside of ef-tests.
 /// See [rustc-hash](https://crates.io/crates/rustc-hash) for more information.
-#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct State {
     pub(crate) classes: FxHashMap<ClassHash, ContractClass>,
     pub(crate) compiled_class_hashes: FxHashMap<ClassHash, CompiledClassHash>,
@@ -41,29 +37,6 @@ impl State {
     /// Helper function allowing to set the nonce of a contract.
     pub fn set_nonce(&mut self, contract_address: ContractAddress, nonce: Nonce) {
         self.nonces.insert(contract_address, nonce);
-    }
-}
-
-impl DumpLoad for State {
-    /// This will serialize the current state, and will save it to a path
-    fn dump_state_to_file(self, path: &Path) -> Result<(), SerializationError> {
-        let serializable_state: SerializableState = self.into();
-
-        let dump = serde_json::to_string(&serializable_state)
-            .map_err(SerializationError::SerdeJsonError)?;
-
-        fs::write(path, dump).map_err(SerializationError::IoError)?;
-
-        Ok(())
-    }
-
-    /// This will read a dump from a file and initialize the state from it
-    fn load_state_from_file(path: &Path) -> Result<Self, SerializationError> {
-        let dump = fs::read(path).unwrap();
-        let serializable_state: SerializableState =
-            serde_json::from_slice(&dump).map_err(SerializationError::SerdeJsonError)?;
-
-        Ok(serializable_state.into())
     }
 }
 
@@ -196,25 +169,12 @@ impl BlockifierStateReader for &mut State {
     }
 }
 
-impl From<SerializableState> for State {
-    fn from(serializable_state: SerializableState) -> Self {
-        Self {
-            classes: serializable_state.classes.clone(),
-            compiled_class_hashes: serializable_state.compiled_classes_hash.clone(),
-            contracts: serializable_state.contracts.clone(),
-            storage: serializable_state.storage.clone(),
-            nonces: serializable_state.nonces.clone(),
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use blockifier::execution::contract_class::ContractClassV0;
 
     use crate::constants::test_constants::{
-        ONE_CLASS_HASH, ONE_COMPILED_CLASS_HASH, ONE_FELT, ONE_PATRICIA, TEST_CONTRACT, TEST_NONCE,
-        TEST_STORAGE_KEY,
+        ONE_CLASS_HASH, ONE_COMPILED_CLASS_HASH, ONE_FELT, ONE_PATRICIA, TEST_CONTRACT,
     };
 
     use super::*;
@@ -318,44 +278,5 @@ mod tests {
 
         // When
         state.get_compiled_class_hash(*ONE_CLASS_HASH).unwrap();
-    }
-
-    #[test]
-    pub fn dump_and_load_state() {
-        let mut state = State::default();
-
-        // setting up entry for state.classes
-        let class_hash = *ONE_CLASS_HASH;
-        let contract_class = include_str!("./test_data/cairo_0/compiled_classes/counter.json");
-        let contract_class: ContractClassV0 =  serde_json::from_str(contract_class).expect("failed to deserialize ContractClass from ./crates/sequencer/test_data/cairo_1/compiled_classes/account.json");
-        let contract_class = ContractClass::V0(contract_class);
-
-        let compiled_class_hash = *ONE_COMPILED_CLASS_HASH;
-        let contract_address = *TEST_CONTRACT;
-        let contract_storage_key: ContractStorageKey = (contract_address, *TEST_STORAGE_KEY);
-        let storage_value = *ONE_FELT;
-        let nonce = *TEST_NONCE;
-
-        state.classes.insert(class_hash, contract_class);
-        state
-            .compiled_class_hashes
-            .insert(class_hash, compiled_class_hash);
-        state.contracts.insert(contract_address, class_hash);
-        state.storage.insert(contract_storage_key, storage_value);
-        state.nonces.insert(contract_address, nonce);
-
-        let temp_file = tempfile::NamedTempFile::new().expect("failed open named temp file");
-        let dump_file_path = temp_file.into_temp_path();
-
-        state
-            .clone()
-            .dump_state_to_file(&dump_file_path)
-            .expect("failed to save dump to file");
-
-        let loaded_state =
-            State::load_state_from_file(&dump_file_path).expect("failed to load state from file");
-        assert_eq!(state, loaded_state);
-
-        dump_file_path.close().expect("failed to close temp file");
     }
 }
